+++ inst/include/epiworld/database-bones.hpp
@@ -163,6 +163,8 @@ public:
      * @return In `get_hist_virus`, the time series of what for each virus.
      * @return In `get_hist_total_date` and `get_hist_virus_date` the
      * corresponding date
+     * @return In `get_active_cases`, the number of active cases (currently infected individuals)
+     * for each virus at each point in time.
      */
     ///@{
     int get_today_total(const std::string & what) const;
@@ -209,6 +211,18 @@ public:
         std::vector< int > & counts,
         bool skip_zeros
     ) const;
+
+    void get_active_cases(
+        std::vector< int > & date,
+        std::vector< int > & virus_id,
+        std::vector< int > & count
+    ) const;
+
+    void get_outbreak_size(
+        std::vector< int > & date,
+        std::vector< int > & virus_id,
+        std::vector< int > & size
+    ) const;
     ///@}
 
     /**
@@ -247,7 +261,9 @@ public:
         std::string fn_transmission,
         std::string fn_transition,
         std::string fn_reproductive_number,
-        std::string fn_generation_time
+        std::string fn_generation_time,
+        std::string fn_active_cases,
+        std::string fn_outbreak_size
         ) const;
 
     /***